+++ Tests/Database.test.ps1
@@ -1,3 +1,4 @@
+#TODO https://github.com/nightroman/MyLiteDB/tree/main/gist/_220616_u5-ReadOnly
 
 Import-Module Ldbc
 
@@ -16,7 +17,7 @@ task NewMemory {
 task NewStream {
 	$stream = [System.IO.MemoryStream]::new()
 
-	$db = New-LiteDatabase $stream
+	$db = New-LiteDatabase -Stream $stream
 	equals $db.GetType() ([LiteDB.LiteDatabase])
 	equals $db.UserVersion 0
 	$db.UserVersion = 26
@@ -24,7 +25,7 @@ task NewStream {
 
 	equals $stream.ToArray().Length 8192
 
-	$db = New-LiteDatabase $stream
+	$db = New-LiteDatabase -Stream $stream
 	equals $db.GetType() ([LiteDB.LiteDatabase])
 	equals $db.UserVersion 26
 	$db.Dispose()
@@ -53,7 +54,7 @@ task UseFile {
 task UseStream {
 	$stream = [System.IO.MemoryStream]::new()
 
-	Use-LiteDatabase $stream {
+	Use-LiteDatabase -Stream $stream {
 		equals $Database.GetType() ([LiteDB.LiteDatabase])
 		equals $Database.UserVersion 0
 		$Database.UserVersion = 26
@@ -61,7 +62,7 @@ task UseStream {
 
 	equals $stream.ToArray().Length 8192
 
-	Use-LiteDatabase $stream {
+	Use-LiteDatabase -Stream $stream {
 		equals $Database.GetType() ([LiteDB.LiteDatabase])
 		equals $Database.UserVersion 26
 	}
@@ -90,12 +91,12 @@ task GetCollectionDoesNotCreate {
 task AutoCommit {
 	$stream = [System.IO.MemoryStream]::new()
 
-	$r = Use-LiteDatabase $stream -Transaction {
+	$r = Use-LiteDatabase -Stream $stream -Transaction {
 		Invoke-LiteCommand 'INSERT INTO test : INT VALUES {p1 : 1}'
 	}
 	equals $r 1
 
-	$r = Use-LiteDatabase $stream {
+	$r = Use-LiteDatabase -Stream $stream {
 		Invoke-LiteCommand 'SELECT $ FROM test'
 	}
 	equals "$r" '{"_id":1,"p1":1}'
@@ -107,7 +108,7 @@ task AutoRollback {
 	$stream = [System.IO.MemoryStream]::new()
 
 	try {
-		Use-LiteDatabase $stream -Transaction {
+		Use-LiteDatabase -Stream $stream -Transaction {
 			$r = Invoke-LiteCommand 'INSERT INTO test : INT VALUES {p1 : 1}'
 			equals $r 1
 
@@ -123,8 +124,77 @@ task AutoRollback {
 		assert ("$_" -like "oops*At*Database.test.ps1*throw 'oops'*")
 	}
 
-	Use-LiteDatabase $stream {
+	Use-LiteDatabase -Stream $stream {
 		$r = Invoke-LiteCommand 'SELECT $ FROM test'
 		equals $r $null
 	}
+}
+
+task ConnectionDirectSharedFails {
+	if ($PSVersionTable.PSVersion.Major -ge 7) {
+		# Method not found: 'Void System.Threading.Mutex..ctor(Boolean, System.String, Boolean ByRef, System.Security.AccessControl.MutexSecurity)'.
+		Write-Warning 'TODO PS7 problem, not new'
+		return
+	}
+
+	remove z.LiteDB
+
+	$Database = New-LiteDatabase z.LiteDB -Connection Direct
+	try
+	{
+		$c1 = Get-LiteCollection c1
+		@{ _id = 1 } | Add-LiteData $c1
+
+		$job = Start-Job {
+			$ErrorActionPreference = 1
+			Import-Module Ldbc
+			Use-LiteDatabase $using:BuildRoot\z.LiteDB -Connection Shared {
+				$c1 = Get-LiteCollection c1
+				Get-LiteData $c1
+			}
+		}
+
+		($r = try { $job | Wait-Job | Receive-Job } catch { "$_" })
+		assert ($r -like 'The process cannot access the file *')
+	}
+	finally {
+		$Database.Dispose()
+	}
+
+	Remove-Item z.LiteDB
+}
+
+task ConnectionSharedDirectWorks {
+	if ($PSVersionTable.PSVersion.Major -ge 7) {
+		# Method not found: 'Void System.Threading.Mutex..ctor(Boolean, System.String, Boolean ByRef, System.Security.AccessControl.MutexSecurity)'.
+		Write-Warning 'TODO PS7 problem, not new'
+		return
+	}
+
+	remove z.LiteDB
+
+	$Database = New-LiteDatabase z.LiteDB -Connection Shared
+	try
+	{
+		$c1 = Get-LiteCollection c1
+		@{ _id = 1 } | Add-LiteData $c1
+
+		$job = Start-Job {
+			$ErrorActionPreference = 1
+			Import-Module Ldbc
+			Use-LiteDatabase $using:BuildRoot\z.LiteDB -Connection Direct {
+				$c1 = Get-LiteCollection c1
+				$r = Get-LiteData $c1
+				"$r"
+			}
+		}
+
+		($r = try { $job | Wait-Job | Receive-Job } catch { "$_" })
+		equals $r '{"_id":1}'
+	}
+	finally {
+		$Database.Dispose()
+	}
+
+	Remove-Item z.LiteDB
 }